He creado una clase DataTypesValidation para ahi insertar todas las funciones de validación. He actualizado el paquete de Testing.
