+++ nbproject/private/private.properties
@@ -1,0 +1,1 @@
+user.properties.file=/home/ferynando7/.netbeans/8.2/build.properties
