+++ src/Forms/NewPatient.java
@@ -6,6 +6,7 @@
 package Forms;
 
 import Logic.ConnectionPostgres;
+import Logic.DataTypesValidation;
 import javax.swing.JFrame;
 
 /**
@@ -17,6 +18,10 @@ public class NewPatient extends javax.swing.JFrame {
     /**
      * Creates new form NewPatient
      */
+    
+    
+    private DataTypesValidation validate = new DataTypesValidation();
+    
     public NewPatient() {
         initComponents();
     }
@@ -284,90 +289,12 @@ public class NewPatient extends javax.swing.JFrame {
     
     */
     public String getTextID(){
-        String id = ID.getText();
-        if (id.length()==10){
-            if (isNumeric(id)){
-                if(codProvCorrect(id) && digVerifCorrect(id)){
-                    return id;
-                }else{
-                    return "Exception 003";
-                }
-            }else{
-                return "Exception 001";
-            }
-        }else{
-            return "Exception 002";
-        }
+        return validate.validateID(ID.getText());
     }
     
     
     
     
-    public String getTextID(String id){
-       
-        if (id.length()==10){
-            if (isNumeric(id)){
-
-                if(codProvCorrect(id) && digVerifCorrect(id) ){
-                    return id;
-                }else{
-                    return "Exception 003";
-                }
-            }else{
-                return "Exception 001";
-            }
-        }else{
-            return "Exception 002";
-        }
-    }
-    
-    
-    private static boolean isNumeric(String cadena){
-        try{
-            Integer.parseInt(cadena);
-            return true;
-        }catch (NumberFormatException nfe){
-            return false;
-        }
-    }
-    
-    private static boolean codProvCorrect(String cadena){
-        String twoFirstDigits = cadena.substring(0,1);
-        int digits = Integer.parseInt(twoFirstDigits);
-        if((digits>0 && digits<25) || digits == 30){
-            return true;
-        }else{
-            return false;
-        }
-    }
-    
-    private static boolean digVerifCorrect(String id){
-        int impar = 0;
-        int par = 1;
-        int sumaImpares=0;
-        int sumaPares=0;
-        int total;
-        
-        for (int i = impar; i < 9;i=i+2){
-            int valueImpar = Character.getNumericValue(id.charAt(i))*2;
-
-            if(valueImpar <=10){ 
-                sumaImpares+= valueImpar;
-            }else{
-                sumaImpares+=valueImpar -9;
-            }
-        }
-        
-        for (int i = par; i<8;i=i+2 ){
-            int valuePar = Character.getNumericValue(id.charAt(i));
-             sumaPares += valuePar;
-        }
-        
-        total = sumaImpares+sumaPares;
-        
-        if(10-total%10 == Character.getNumericValue(id.charAt(9))) return true;
-        else return false;
-    }
 
     
     /**
